change authorization method

+++ application/modules/users/controllers/Users.php
@@ -9,7 +9,7 @@
 
 class Users extends MX_Controller
 {
-	private $is_login;
+	private $_access;
 
 	function __construct()
 	{
@@ -17,7 +17,7 @@ class Users extends MX_Controller
 		$this->load->module('auth');
 		$this->load->model('UsersModel');
 		$this->load->helper('date');
-		$this->is_login = $this->auth->is_login();
+		$this->_access = $this->auth->privileges_read('users_browse');
 	}
 
 	/**
@@ -25,36 +25,19 @@ class Users extends MX_Controller
 	 */
 	public function index()
 	{
-		$this->browse();
-	}
-
-	public function test()
-	{
-		/**
-		 * this function must be delete before deploy
-		 */
-		$this->show_interface('test','');
-
+		$this->_browse();
 	}
 
 	/**
 	 * browse method
 	 * get all user's data from persistence storage
 	 */
-	public function browse()
+	public function _browse()
 	{
-		if($this->is_login)
+		if($this->_access)
 		{
-			$access = $this->auth->privileges_read('users_browse');
-			if($access)
-			{
-				$data['users_data'] = $this->UsersModel->browse();
-				$this->show_interface('browse', $data);
-			}
-			else
-			{
-				$this->show_interface('Unauthorize', '');
-			}
+			$data['users_data'] = $this->UsersModel->browse();
+			$this->_show_interface('browse', $data);
 		}
 		else
 		{
@@ -68,19 +51,11 @@ class Users extends MX_Controller
 	 */
 	public function read($email)
 	{
-		if($this->is_login)
+		$this->_access = $this->auth->privileges_read('users_read');
+		if($this->_access)
 		{
-			$access = $this->auth->privileges_read('users_read');
-			if($access)
-			{
-				$data['user_data'] = $this->UsersModel->read($email);
-				$this->show_interface('read', $data);
-			}
-			else
-			{
-				$this->show_interface('Unauthorize', '');
-			}
-
+			$data['user_data'] = $this->UsersModel->read($email);
+			$this->_show_interface('read', $data);
 		}
 		else
 		{
@@ -94,52 +69,44 @@ class Users extends MX_Controller
 	 */
 	public function edit($id)
 	{
-		if($this->is_login)
+		$this->_access = $this->auth->privileges_read('users_edit');
+		if($this->_access)
 		{
-			$access = $this->auth->privileges_read('users_edit');
-			if($access)
-			{
-				$datestring = '%Y-%m-%d %H:%i:%s';
-				if (substr($id, 0, 2) !== 'IU') {
-					$data['user_data'] = $this->UsersModel->read($id);
-					$this->show_interface('edit', $data);
-				}
-				else
-				{
-					$data = array(
-							'users' => array(
-									'email' => $this->input->post('email'), 
-									'password' => $this->input->post('password'),
-									'jabatan' => $this->input->post('jabatan'),
-									'name_full' => $this->input->post('nama'),
-									'ktp_no' => $this->input->post('ktp_no'),
-									'ktp_img' => $this->input->post('ktp_img'),
-									'alamat' => $this->input->post('alamat'),
-									'telepon_primer' => $this->input->post('telepon_primer'),
-									'rekening_no' => $this->input->post('rekening_no'),
-									'rekening_bank' => $this->input->post('rekening_bank'),
-									'agama' => $this->input->post('agama'), 
-									'status' => $this->input->post('status'),
-									'anak' => $this->input->post('anak'),
-									'photo' => $this->input->post('photo')
-								),
-							'details' => array(
-									'telepon_sekunder' => ($this->input->post('telepon_sekunder') === '' ? 'null' : $this->input->post('telepon_sekunder')),
-									'telepon_pin' => ($this->input->post('telepon_pin') === '' ? 'null' : $this->input->post('telepon_pin')),
-									'telepon_whatsapp' => ($this->input->post('telepon_whatsapp') === '' ? 'null' : $this->input->post('telepon_whatsapp')),
-									'npwp_no' => ($this->input->post('npwp_no') === '' ? 'null' : $this->input->post('npwp_no')),
-									'npwp_img' => ($this->input->post('npwp_img') === '' ? 'null' : $this->input->post('npwp_img')),
-									'created_at' => mdate($datestring, time()),
-									'edited_at' => mdate($datestring, time())
-								)
-						);
-					$this->UsersModel->edit($id, $data['users'], $data['details']);
-					redirect(base_url("/users/read/".$data['users']['email']));
-				}
+			$datestring = '%Y-%m-%d %H:%i:%s';
+			if (substr($id, 0, 2) !== 'IU') {
+				$data['user_data'] = $this->UsersModel->read($id);
+				$this->_show_interface('edit', $data);
 			}
 			else
 			{
-				$this->show_interface('Unauthorize', '');
+				$data = array(
+						'users' => array(
+								'email' => $this->input->post('email'), 
+								'password' => $this->input->post('password'),
+								'jabatan' => $this->input->post('jabatan'),
+								'name_full' => $this->input->post('nama'),
+								'ktp_no' => $this->input->post('ktp_no'),
+								'ktp_img' => $this->input->post('ktp_img'),
+								'alamat' => $this->input->post('alamat'),
+								'telepon_primer' => $this->input->post('telepon_primer'),
+								'rekening_no' => $this->input->post('rekening_no'),
+								'rekening_bank' => $this->input->post('rekening_bank'),
+								'agama' => $this->input->post('agama'), 
+								'status' => $this->input->post('status'),
+								'anak' => $this->input->post('anak'),
+								'photo' => $this->input->post('photo')
+							),
+						'details' => array(
+								'telepon_sekunder' => ($this->input->post('telepon_sekunder') === '' ? 'null' : $this->input->post('telepon_sekunder')),
+								'telepon_pin' => ($this->input->post('telepon_pin') === '' ? 'null' : $this->input->post('telepon_pin')),
+								'telepon_whatsapp' => ($this->input->post('telepon_whatsapp') === '' ? 'null' : $this->input->post('telepon_whatsapp')),
+								'npwp_no' => ($this->input->post('npwp_no') === '' ? 'null' : $this->input->post('npwp_no')),
+								'npwp_img' => ($this->input->post('npwp_img') === '' ? 'null' : $this->input->post('npwp_img')),
+								'created_at' => mdate($datestring, time()),
+								'edited_at' => mdate($datestring, time())
+							)
+					);
+				($this->UsersModel->edit($id, $data['users'], $data['details']) ? redirect(base_url("/users/read/".$data['users']['email'])) : redirect(base_url()));
 			}
 		}
 		else
@@ -154,170 +121,162 @@ class Users extends MX_Controller
 	 */
 	public function add()
 	{
-		if($this->is_login)
+		$this->_access =$this->auth->privileges_read('users_add');
+		if($this->_access)
 		{
-			$access =$this->auth->privileges_read('users_add');
 			$datestring = '%Y-%m-%d %H:%i:%s';
 			$id = $this->UsersModel->idGen();
-			if($access)
+			$upload_conf = array(
+				'allowed_types' => 'jpg|png|jpeg',
+				'max_size' => '10000',
+				'file_name' => $id,
+				'overwrite' => TRUE,
+			);
+			$resize_conf = array(
+				'image_library' => 'gd2',
+				'create_thumb' => TRUE,
+				'maintain_ratio' => TRUE,
+				'width' => 600,
+			);
+
+			if(!empty($this->input->post('email')))
 			{
-				$upload_conf = array(
-					'allowed_types' => 'jpg|png|jpeg',
-					'max_size' => '10000',
-					'file_name' => $id,
-					'overwrite' => TRUE,
-					);
-				$resize_conf = array(
-						'image_library' => 'gd2',
-						'create_thumb' => TRUE,
-						'maintain_ratio' => TRUE,
-						'width' => 600,
-					);
+				$upload_conf['upload_path'] = './assets/ktp';
+				$this->load->library('upload', $upload_conf);
+				$this->load->library('image_lib',$resize_conf);
 
-				if(!empty($this->input->post('email')))
+				if($this->upload->do_upload('ktp_img'))
 				{
-					$upload_conf['upload_path'] = './assets/ktp';
-					$this->load->library('upload', $upload_conf);
-					$this->load->library('image_lib',$resize_conf);
+					$ktp = ''.$id.''.$this->upload->data('file_ext');
+					chmod($this->upload->data('full_path'), 0777);
 
-					if($this->upload->do_upload('ktp_img'))
+					$upload_conf['upload_path'] = './assets/photo/';
+					$this->upload->initialize($upload_conf);
+					if($this->upload->do_upload('user_img'))
 					{
-						$ktp = ''.$id.''.$this->upload->data('file_ext');
 						chmod($this->upload->data('full_path'), 0777);
 
-						$upload_conf['upload_path'] = './assets/photo/';
-						$this->upload->initialize($upload_conf);
-						if($this->upload->do_upload('user_img'))
+						$resize_conf['source_image'] = $this->upload->data('full_path');
+						$this->image_lib->initialize($resize_conf);
+						if($this->image_lib->resize())
 						{
-							chmod($this->upload->data('full_path'), 0777);
-
-							$resize_conf['source_image'] = $this->upload->data('full_path');
-							$this->image_lib->initialize($resize_conf);
-							if($this->image_lib->resize())
+							$img = $this->upload->data('file_name');
+							$img_thumb = ''.$id.'_thumb'.$this->upload->data('file_ext');
+							if(!empty($this->input->post('npwp_img')))
 							{
-								$img = $this->upload->data('file_name');
-								$img_thumb = ''.$id.'_thumb'.$this->upload->data('file_ext');
-								if(!empty($this->input->post('npwp_img')))
-								{
-									$upload_conf['upload_path'] = './assets/npwp';
-									$this->upload->initialize($upload_conf);
-									$this->upload->do_upload('npwp_img');
-									$npwp = ''.$id.''.$this->upload->data('file_ext');
-								}
+								$upload_conf['upload_path'] = './assets/npwp';
+								$this->upload->initialize($upload_conf);
+								$this->upload->do_upload('npwp_img');
+								$npwp = ''.$id.''.$this->upload->data('file_ext');
+							}
 
-								$data = array(
-								'users' => array(
-											'id' => $id,
-											'email' => $this->input->post('email'),
-											'password' => $this->input->post('password'),
-											'jabatan' => $this->input->post('jabatan'),
-											'name_full' => $this->input->post('nama'),
-											'ktp_no' => $this->input->post('ktp_no'),
-											'ktp_img' => $ktp,
-											'alamat' => $this->input->post('alamat'),
-											'telepon_primer' => $this->input->post('telepon_primer'),
-											'rekening_no' => $this->input->post('rekening_no'),
-											'rekening_bank' => $this->input->post('rekening_bank'),
-											'agama' => $this->input->post('agama'),
-											'status' => $this->input->post('status'),
-											'anak' => $this->input->post('anak'),
-											'photo' => $img,
-											'thumbnail' => $img_thumb,
-										),
-								'details' => array(
-											'id' => $id,
-											'telepon_sekunder' => ($this->input->post('telepon_sekunder') === '' ? 'null' : $this->input->post('telepon_sekunder')),
-											'telepon_pin' => ($this->input->post('telepon_pin') === '' ? 'null' : $this->input->post('telepon_pin')),
-											'telepon_whatsapp' => ($this->input->post('telepon_whatsapp') === '' ? 'null' : $this->input->post('telepon_whatsapp')),
-											'npwp_no' => ($this->input->post('npwp_no') === '' ? 'null' : $this->input->post('npwp_no')),
-											'npwp_img' => ($npwp === '' ? 'null' : $npwp),
-											'created_at' => mdate($datestring, time()),
-											'edited_at' => mdate($datestring, time())
-										),
-								'privileges' => array(
-											'id' => $id,
-											'privileges_browse' => 0,
-											'privileges_read' => 0,
-											'privileges_edit' => 0,
-											'users_browse' => 0,
-											'users_read' => 0,
-											'users_edit' => 0,
-											'users_add' => 0,
-											'users_delete' => 0,
-											'standard_browse' => 0,
-											'standard_edit' => 0,
-											'standard_add' => 0,
-											'standard_delete' => 0,
-											'contract_browse' => 0,
-											'contract_read' => 0,
-											'contract_edit' => 0,
-											'contract_add' => 0,
-											'contract_delete' => 0,
-											'ring_browse' => 0,
-											'ring_read' => 0,
-											'ring_edit' => 0,
-											'ring_add' => 0,
-											'ring_delete' => 0,
-											'ts_browse' => 0,
-											'ts_read' => 0,
-											'ts_edit' => 0,
-											'ts_add' => 0,
-											'ts_delete' => 0,
-											'breeder_browse' => 0,
-											'breeder_read' => 0,
-											'breeder_edit' => 0,
-											'breeder_add' => 0,
-											'breeder_delete' => 0,
-											'supplier_browse' => 0,
-											'supplier_read' => 0,
-											'supplier_edit' => 0,
-											'supplier_add' => 0,
-											'supplier_delete' => 0,
-											'supplier_prod_browse' => 0,
-											'supplier_prod_read' => 0,
-											'supplier_prod_edit' => 0,
-											'supplier_prod_add' => 0,
-											'supplier_prod_delete' => 0,
-											'buyer_browse' => 0,
-											'buyer_read' => 0,
-											'buyer_edit' => 0,
-											'buyer_add' => 0,
-											'buyer_delete' => 0,
-											'breeder_score_browse' => 0,
-											'breeder_score_read' => 0,
-											'breeder_score_edit' => 0,
-											'breeder_score_add' => 0,
-											'breeder_score_delete' => 0
-										)
-									);
-							
-								($this->UsersModel->add($data['users'], $data['details'], $data['privileges'])) ? redirect(base_url('/users/read/'.$this->input->post('email'))) : redirect(base_url('/users/add')) ;
-							}
-							else
-							{
-								$data['error'] = $this->image_lib->display_errors();
-								$this->load->view('test',$data);
-							}
+							$data = array(
+							'users' => array(
+										'id' => $id,
+										'email' => $this->input->post('email'),
+										'password' => $this->input->post('password'),
+										'jabatan' => $this->input->post('jabatan'),
+										'name_full' => $this->input->post('nama'),
+										'ktp_no' => $this->input->post('ktp_no'),
+										'ktp_img' => $ktp,
+										'alamat' => $this->input->post('alamat'),
+										'telepon_primer' => $this->input->post('telepon_primer'),
+										'rekening_no' => $this->input->post('rekening_no'),
+										'rekening_bank' => $this->input->post('rekening_bank'),
+										'agama' => $this->input->post('agama'),
+										'status' => $this->input->post('status'),
+										'anak' => $this->input->post('anak'),
+										'photo' => $img,
+										'thumbnail' => $img_thumb,
+									),
+							'details' => array(
+										'id' => $id,
+										'telepon_sekunder' => ($this->input->post('telepon_sekunder') === '' ? 'null' : $this->input->post('telepon_sekunder')),
+										'telepon_pin' => ($this->input->post('telepon_pin') === '' ? 'null' : $this->input->post('telepon_pin')),
+										'telepon_whatsapp' => ($this->input->post('telepon_whatsapp') === '' ? 'null' : $this->input->post('telepon_whatsapp')),
+										'npwp_no' => ($this->input->post('npwp_no') === '' ? 'null' : $this->input->post('npwp_no')),
+										'npwp_img' => ($npwp === '' ? 'null' : $npwp),
+										'created_at' => mdate($datestring, time()),
+										'edited_at' => mdate($datestring, time())
+									),
+							'privileges' => array(
+										'id' => $id,
+										'privileges_browse' => 0,
+										'privileges_read' => 0,
+										'privileges_edit' => 0,
+										'users_browse' => 0,
+										'users_read' => 0,
+										'users_edit' => 0,
+										'users_add' => 0,
+										'users_delete' => 0,
+										'standard_browse' => 0,
+										'standard_edit' => 0,
+										'standard_add' => 0,
+										'standard_delete' => 0,
+										'contract_browse' => 0,
+										'contract_read' => 0,
+										'contract_edit' => 0,
+										'contract_add' => 0,
+										'contract_delete' => 0,
+										'ring_browse' => 0,
+										'ring_read' => 0,
+										'ring_edit' => 0,
+										'ring_add' => 0,
+										'ring_delete' => 0,
+										'ts_browse' => 0,
+										'ts_read' => 0,
+										'ts_edit' => 0,
+										'ts_add' => 0,
+										'ts_delete' => 0,
+										'breeder_browse' => 0,
+										'breeder_read' => 0,
+										'breeder_edit' => 0,
+										'breeder_add' => 0,
+										'breeder_delete' => 0,
+										'supplier_browse' => 0,
+										'supplier_read' => 0,
+										'supplier_edit' => 0,
+										'supplier_add' => 0,
+										'supplier_delete' => 0,
+										'supplier_prod_browse' => 0,
+										'supplier_prod_read' => 0,
+										'supplier_prod_edit' => 0,
+										'supplier_prod_add' => 0,
+										'supplier_prod_delete' => 0,
+										'buyer_browse' => 0,
+										'buyer_read' => 0,
+										'buyer_edit' => 0,
+										'buyer_add' => 0,
+										'buyer_delete' => 0,
+										'breeder_score_browse' => 0,
+										'breeder_score_read' => 0,
+										'breeder_score_edit' => 0,
+										'breeder_score_add' => 0,
+										'breeder_score_delete' => 0
+									)
+								);
+						
+							($this->UsersModel->add($data['users'], $data['details'], $data['privileges'])) ? redirect(base_url('/users/read/'.$this->input->post('email'))) : redirect(base_url('/users/add')) ;
+						}
+						else
+						{
+							$data['error'] = $this->image_lib->display_errors();
+							$this->load->view('test',$data);
 						}
 					}
-					
-					else
-					{
-						$error = $this->upload->display_error();
-						$this->load->view('test',$error);
-					}
 				}
-
+				
 				else
 				{
-					$data['id'] = $this->UsersModel->idGen();
-					$this->show_interface('add', $data);
+					$error = $this->upload->display_error();
+					$this->load->view('test',$error);
 				}
 			}
 
 			else
 			{
-				$this->show_interface('Unauthorize', '');
+				$data['id'] = $this->UsersModel->idGen();
+				$this->_show_interface('add', $data);
 			}
 		}
 
@@ -333,26 +292,16 @@ class Users extends MX_Controller
 	 */
 	public function delete($id)
 	{
-		if($this->is_login)
+		$this->_access = $this->auth->privileges_read('users_delete');
+		if($this->_access && !empty($id))
 		{
-			$access = $this->auth->privileges_read('users_delete');
-			if($access)
+			if($this->UsersModel->delete($id, FALSE))
 			{
-				if (!empty($id))
-				{
-					if($this->UsersModel->delete($id, FALSE))
-					{
-						$id === $this->session->userdata ? redirect(base_url()) : redirect(base_url('/users'));
-					}
-				}
-				else
-				{
-					redirect(base_url('/users'));
-				}
+				($id === $this->session->userdata ? redirect(base_url('auth/logout')) : redirect(base_url('/users')));
 			}
 			else
 			{
-				$this->show_interface('Unauthorize', '');
+				redirect(base_url());
 			}
 		}
 		else
@@ -363,10 +312,10 @@ class Users extends MX_Controller
 	}
 
 	/**
-	 * show_interface method
+	 * _show_interface method
 	 * load user interface page which contain the $data
 	 */
-	private function show_interface($page, $data)
+	function _show_interface($page, $data)
 	{
 		if(!empty($page && $data))
 		{
